requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
django-filter>=23.5
